+++ authentication/tests/test_models.py
@@ -1,5 +1,6 @@
 import datetime
 
+from api import conf
 from api.models import Permission
 
 
@@ -7,7 +8,7 @@ def test_permission_serializer():
     test_permission = Permission(
         oauthIssuer="https://api.example.com/issuer",
         client="https://directory.core.pilot.trust.ib1.org/member/28364528",
-        license="https://registry.core.sandbox.trust.ib1.org/scheme/perseus/license/energy-consumption-edp-cap/2026-03-12",
+        license=conf.ENERGY_CONSUMPTION_LICENSE_URL,
         account="6qIO3KZx0Q",
         lastGranted=datetime.datetime.now(datetime.timezone.utc),
         expires="2025-03-31T23:30Z",